+++ tests/test_make_and_undo_move.cpp
@@ -15,7 +15,7 @@ static SearchPosition19 emptyPos()
 // current position, then apply.
 static void doMove(SearchPosition19& pos, int col, int row, Color color)
 {
-    pos.makeMove(col, row, color, pos.buildMoveHash(EvaluatedMove{t_cell{col, row}, 0, true, {}}, color));
+    pos.makeMove(col, row, color, pos.buildMoveHash(EvaluatedMove{t_cell{static_cast<int_fast16_t>(col), static_cast<int_fast16_t>(row)}, 0, true, {}}, color));
 }
 
 TEST_CASE("make/undo single move restores hash")
